backend/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
cached-property==1.5.2
distlib==0.3.8
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==5.0.2
Django==4.2.11
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==1.1.0
django-rest-framework==0.1.0
django-url-filter==0.3.15
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-filters==0.11.1
enum-compat==0.0.3
environs==9.5.0
filelock==3.13.1
gunicorn==20.1.0
marshmallow==3.21.1
packaging==24.0
platformdirs==4.2.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
virtualenv==20.25.1
whitenoise==6.3.0
